# Conflicts:
#	app/src/main/java/org/p2p/wallet/common/ui/widget/ProgressButton.kt
#	app/src/main/java/org/p2p/wallet/utils/ContextExtensions.kt
#	app/src/main/res/drawable/bg_blue_disabled_rounded.xml
#	app/src/main/res/values/colors.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/org/p2p/wallet/rpc/repository/RpcRepository.kt
@@ -35,7 +35,7 @@ interface RpcRepository {
     ): List<ProgramAccount>
 
     suspend fun getBalance(account: String): Long
-    suspend fun getBalances(accounts: List<String>): List<Pair<String, Long>>
+    suspend fun getBalances(accounts: List<String>): List<Pair<String, BigInteger>>
     suspend fun getTokenAccountsByOwner(owner: String): TokenAccounts
     suspend fun getMinimumBalanceForRentExemption(dataLength: Long): Long
     suspend fun getMultipleAccounts(publicKeys: List<PublicKey>): MultipleAccountsInfo
